SASchallengeproject/requirements.txt
Django==5.2.7
python-dotenv==1.0.0
requests==2.31.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
